backend/requirements.txt
streamlit
pandas
plotly
autogluon.tabular[all]
openpyxl
pyyaml
pip-tools
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
dotenv
python-multipart
XlsxWriter